+++ backend/core-api/src/modules/contacts/graphql/schemas/customer.ts
@@ -55,7 +55,7 @@ export const types = `
   type CustomersListResponse {
     list: [Customer],
     pageInfo: PageInfo
-    totalCount: Float,
+    totalCount: Int,
   }
 
 `;
